feat: add interface in enthusiast-common, and small refactor
