+++ plugins/enthusiast-common/enthusiast_common/__init__.py
@@ -1,5 +1,6 @@
 from .interfaces import DocumentSourcePlugin as DocumentSourcePlugin
 from .interfaces import ProductSourcePlugin as ProductSourcePlugin
 from .interfaces import EmbeddingProvider as EmbeddingProvider
+from .interfaces import CustomTool as CustomTool
 from .structures import DocumentDetails as DocumentDetails
 from .structures import ProductDetails as ProductDetails
