+++ server/agent/tools/__init__.py
@@ -1,4 +1,3 @@
 from .create_answer_tool import CreateAnswerTool
-from .duck_tool import DuckTool
 
-__all__ = ['CreateAnswerTool', 'DuckTool']
+__all__ = ['CreateAnswerTool']
