+++ server/agent/tools/registry.py
